core: Moved obstrvable struct definitions to sep file
